requirements.txt
# Core RAG
llama-index
llama-index-llms-groq
llama-index-embeddings-huggingface
llama-index-readers-file
llama-index-vector-stores-faiss
llama-index-postprocessor-cohere-rerank
llama-index-retrievers-bm25

# Performance & UI
streamlit
python-dotenv
pydantic-settings
faiss-cpu
rank-bm25

# Security
pip-audit

